+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## 1.0.2
+## 1.0.3
 
 - Dialog Height Issue Resolved.
+- Horizontal Drag Vertex improved
+- Custom Button Title